Updating numeric conversions
When dealing with JSON.Net integer JValues as double, it is not
necessary to first convert to (int) before converting to double.  In
fact, this causes problems when dealing with integer outside of range
for Int32.

The reverse is not true when working with JSON.Net double values as
integer.  Therefore the peer method AsInteger() still requires explicit
cast to double then cast to int.