Rough outline of Context Usage
Provides the base outline of working with multiple managed objects,
allowing user code to deal directly with NSManagedObjects.

The base idea is the store opts out of user changes.  We will provide a
new MSManagedObjectObserver that will do the typical right thing on a
context save.  (Generating the sync table updates/inserts/deletes)