Unit test and potential fix for #612.
The issue here is that when performing sequential updates or an insert followed by an update on the server, the server barfs on the second update because the version provided by the server after the first operation isn't saved and therefore isn't sent to the server when attempting the second operation.

This change adds a unit test for this case and a potential fix. All unit tests passing with this change.